requirements.txt
appdirs==1.4.4
arrow==1.1.0
asgiref==3.3.4
beautifulsoup4==4.9.3
binaryornot==0.4.4
black==21.5b1
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cookiecutter==1.7.2
coverage==5.5
crispy-tailwind==0.5.0
decorator==4.4.2
Django==3.2.2
django-cleanup==5.2.0
django-crispy-forms==1.11.2
django-debug-toolbar==3.2.1
django-dynamic-formsets==0.0.8
django-environ==0.4.5
django-fieldsets-with-inlines==0.6
django-ical==1.7.3
django-loginas==0.3.9
django-nested-admin==3.3.3
django-phonenumber-field==5.1.0
django-recurrence==1.10.3
django-tailwind==2.2.0
django-tinymce==3.3.0
django-unicorn==0.27.2
flake8==3.9.2
google-api-core==1.26.3
google-api-python-client==2.3.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
icalendar==4.0.7
idna==2.10
Jinja2==2.11.3
jinja2-time==0.2.0
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.0
orjson==3.5.2
packaging==20.9
pathspec==0.8.1
phonenumberslite==8.12.22
pkg_resources==0.0.0
poyo==0.5.0
protobuf==3.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-monkey-business==1.0.0
python-slugify==5.0.2
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
shortuuid==1.0.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
